requirements.txt
fastapi
psycopg2-binary
sqlmodel
python-dotenv
cloudinary
uvicorn[standar]